deepagents/maaya/pyproject.toml
[project]
name = "maaya"
version = "0.1.0"
description = "Maaya — Multi-agent full-stack development orchestrator"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.1.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "aiofiles>=23.0.0",
    "rich>=13.0.0",
]

[project.scripts]
maaya = "cli:main"
maaya-server = "server.main:run"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
